Move addRelation method to EloquentModel
