+++ src/Processor/RelationProcessor.php
@@ -4,24 +4,13 @@ namespace Krlove\EloquentModelGenerator\Processor;
 
 use Doctrine\DBAL\Schema\Table;
 use Illuminate\Database\DatabaseManager;
-use Illuminate\Database\Eloquent\Relations\BelongsTo as EloquentBelongsTo;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany as EloquentBelongsToMany;
-use Illuminate\Database\Eloquent\Relations\HasMany as EloquentHasMany;
-use Illuminate\Database\Eloquent\Relations\HasOne as EloquentHasOne;
-use Illuminate\Support\Str;
-use Krlove\CodeGenerator\Model\DocBlockModel;
-use Krlove\CodeGenerator\Model\MethodModel;
-use Krlove\CodeGenerator\Model\VirtualPropertyModel;
 use Krlove\EloquentModelGenerator\Config\Config;
-use Krlove\EloquentModelGenerator\Exception\GeneratorException;
-use Krlove\EloquentModelGenerator\Helper\EmgHelper;
 use Krlove\EloquentModelGenerator\Helper\Prefix;
 use Krlove\EloquentModelGenerator\Model\BelongsTo;
 use Krlove\EloquentModelGenerator\Model\BelongsToMany;
 use Krlove\EloquentModelGenerator\Model\EloquentModel;
 use Krlove\EloquentModelGenerator\Model\HasMany;
 use Krlove\EloquentModelGenerator\Model\HasOne;
-use Krlove\EloquentModelGenerator\Model\Relation;
 
 class RelationProcessor implements ProcessorInterface
 {
@@ -45,7 +34,7 @@ class RelationProcessor implements ProcessorInterface
                 $tableForeignKey->getLocalColumns()[0],
                 $tableForeignColumns[0]
             );
-            $this->addRelation($model, $relation);
+            $model->addRelation($relation);
         }
 
         $tables = $schemaManager->listTables();
@@ -74,7 +63,7 @@ class RelationProcessor implements ProcessorInterface
                             $localColumns[0],
                             $secondForeignKey->getLocalColumns()[0]
                         );
-                        $this->addRelation($model, $relation);
+                        $model->addRelation($relation);
 
                         break;
                     } else {
@@ -88,7 +77,7 @@ class RelationProcessor implements ProcessorInterface
                             $relation = new HasMany($tableName, $foreignColumn, $localColumn);
                         }
 
-                        $this->addRelation($model, $relation);
+                        $model->addRelation($relation);
                     }
                 }
             }
@@ -114,119 +103,5 @@ class RelationProcessor implements ProcessorInterface
         }
 
         return false;
-    }
-
-    protected function addRelation(EloquentModel $model, Relation $relation): void
-    {
-        $relationClass = EmgHelper::getClassNameByTableName($relation->getTableName());
-        if ($relation instanceof HasOne) {
-            $name = Str::singular(Str::camel($relation->getTableName()));
-            $docBlock = sprintf('@return \%s', EloquentHasOne::class);
-
-            $virtualPropertyType = $relationClass;
-        } elseif ($relation instanceof HasMany) {
-            $name = Str::plural(Str::camel($relation->getTableName()));
-            $docBlock = sprintf('@return \%s', EloquentHasMany::class);
-
-            $virtualPropertyType = sprintf('%s[]', $relationClass);
-        } elseif ($relation instanceof BelongsTo) {
-            $name = Str::singular(Str::camel($relation->getTableName()));
-            $docBlock = sprintf('@return \%s', EloquentBelongsTo::class);
-
-            $virtualPropertyType = $relationClass;
-        } elseif ($relation instanceof BelongsToMany) {
-            $name = Str::plural(Str::camel($relation->getTableName()));
-            $docBlock = sprintf('@return \%s', EloquentBelongsToMany::class);
-
-            $virtualPropertyType = sprintf('%s[]', $relationClass);
-        } else {
-            throw new GeneratorException('Relation not supported');
-        }
-
-        $method = new MethodModel($name);
-        $method->setBody($this->createMethodBody($model, $relation));
-        $method->setDocBlock(new DocBlockModel($docBlock));
-
-        $model->addMethod($method);
-        $model->addProperty(new VirtualPropertyModel($name, $virtualPropertyType));
-    }
-
-    protected function createMethodBody(EloquentModel $model, Relation $relation): string
-    {
-        $reflectionObject = new \ReflectionObject($relation);
-        $name = Str::camel($reflectionObject->getShortName());
-
-        $arguments = [
-            $model->getNamespace()->getNamespace() . '\\' . EmgHelper::getClassNameByTableName($relation->getTableName())
-        ];
-
-        if ($relation instanceof BelongsToMany) {
-            $defaultJoinTableName = EmgHelper::getDefaultJoinTableName(
-                $model->getTableName(),
-                $relation->getTableName()
-            );
-            $joinTableName = $relation->getJoinTable() === $defaultJoinTableName
-                ? null
-                : $relation->getJoinTable();
-            $arguments[] = $joinTableName;
-
-            $arguments[] = $this->resolveArgument(
-                $relation->getForeignColumnName(),
-                EmgHelper::getDefaultForeignColumnName($model->getTableName())
-            );
-            $arguments[] = $this->resolveArgument(
-                $relation->getLocalColumnName(),
-                EmgHelper::getDefaultForeignColumnName($relation->getTableName())
-            );
-        } elseif ($relation instanceof HasMany) {
-            $arguments[] = $this->resolveArgument(
-                $relation->getForeignColumnName(),
-                EmgHelper::getDefaultForeignColumnName($model->getTableName())
-            );
-            $arguments[] = $this->resolveArgument(
-                $relation->getLocalColumnName(),
-                EmgHelper::DEFAULT_PRIMARY_KEY
-            );
-        } else {
-            $arguments[] = $this->resolveArgument(
-                $relation->getForeignColumnName(),
-                EmgHelper::getDefaultForeignColumnName($relation->getTableName())
-            );
-            $arguments[] = $this->resolveArgument(
-                $relation->getLocalColumnName(),
-                EmgHelper::DEFAULT_PRIMARY_KEY
-            );
-        }
-
-        return sprintf('return $this->%s(%s);', $name, $this->prepareArguments($arguments));
-    }
-
-    protected function prepareArguments(array $array): string
-    {
-        $array = array_reverse($array);
-        $milestone = false;
-        foreach ($array as $key => &$item) {
-            if (!$milestone) {
-                if (!is_string($item)) {
-                    unset($array[$key]);
-                } else {
-                    $milestone = true;
-                }
-            } else {
-                if ($item === null) {
-                    $item = 'null';
-
-                    continue;
-                }
-            }
-            $item = sprintf("'%s'", $item);
-        }
-
-        return implode(', ', array_reverse($array));
-    }
-
-    protected function resolveArgument(string $actual, string $default): ?string
-    {
-        return $actual === $default ? null : $actual;
     }
 }
